+++ packages/bentocache/tests/drivers/orchid/postgres.spec.ts
@@ -2,6 +2,7 @@ import { test } from '@japa/runner'
 import { createDb } from 'orchid-orm'
 
 import { createOrchidStore } from './helpers.js'
+import { POSTGRES_CREDENTIALS } from '../../helpers/index.js'
 import { registerCacheDriverTestSuite } from '../../../src/test_suite.js'
 
 test.group('Orchid | Postgres driver', (group) => {
@@ -10,7 +11,7 @@ test.group('Orchid | Postgres driver', (group) => {
     group,
     supportsMilliseconds: false,
     createDriver: (options) => {
-      const db = createDb({ user: 'postgres', password: 'postgres' })
+      const db = createDb({ ...POSTGRES_CREDENTIALS })
 
       return createOrchidStore({ connection: db, prefix: 'japa', ...options })
     },